+++ src/test/kotlin/com/atlassian/performance/tools/jiraactions/api/scenario/JiraCoreScenarioIT.kt
@@ -9,8 +9,8 @@ import com.atlassian.performance.tools.jiraactions.api.measure.output.Collection
 import com.atlassian.performance.tools.jiraactions.api.memories.User
 import com.atlassian.performance.tools.jiraactions.api.memories.UserMemory
 import com.atlassian.performance.tools.jiraactions.api.page.isElementPresent
-import com.atlassian.performance.tools.jiraactions.api.page.wait
 import com.atlassian.performance.tools.jiraactions.api.w3c.DisabledW3cPerformanceTimeline
+import com.atlassian.performance.tools.jiraactions.api.webdriver.sendKeysAndValidate
 import org.apache.logging.log4j.LogManager
 import org.apache.logging.log4j.Logger
 import org.assertj.core.api.Assertions
@@ -18,7 +18,6 @@ import org.junit.Test
 import org.openqa.selenium.By
 import org.openqa.selenium.OutputType
 import org.openqa.selenium.remote.RemoteWebDriver
-import org.openqa.selenium.support.ui.ExpectedConditions
 import java.nio.file.Paths
 import java.time.Clock
 import java.util.*
@@ -130,8 +129,8 @@ class JiraCoreScenarioIT {
     }
 
     private fun addBackupService(driver: RemoteWebDriver) {
-        sendKeys(driver, By.id("serviceName"), "another backup")
-        sendKeys(driver, By.id("serviceClass"), "com.atlassian.jira.service.services.export.ExportService")
+        driver.findElement(By.id("serviceName")).sendKeysAndValidate(driver, "another backup")
+        driver.findElement(By.id("serviceClass")).sendKeysAndValidate(driver, "com.atlassian.jira.service.services.export.ExportService")
 
         driver.findElementById("addservice_submit").click()
 
@@ -142,21 +141,6 @@ class JiraCoreScenarioIT {
             // we can remove this try/catch once we are sure it's stable
             println(driver.getScreenshotAs(OutputType.BASE64))
             throw e
-        }
-    }
-
-    private fun sendKeys(driver: RemoteWebDriver, locator: By, text: String) {
-        val element = driver.wait(ExpectedConditions.elementToBeClickable(locator))
-        element.click()
-
-        // It's hard to say when the keys can be sent. They seem to randomly get lost.
-        for (i in 0..10) {
-            element.sendKeys(text)
-            val currentValue = element.getAttribute("value")
-            if (currentValue == text) {
-                break
-            }
-            Thread.sleep(500)
         }
     }
 }