+++ CHANGELOG.md
@@ -26,6 +26,9 @@ Dropping a requirement of a major version of a dependency is a new contract.
 ## [Unreleased]
 [Unreleased]: https://github.com/atlassian/jira-actions/compare/release-3.10.0...master
 
+### Added
+- Hardened sendKeys() method: ElementUtils#sendKeysAndValidate
+
 ### Fixed
 - `CreateIssueAction` should now properly fill in issue data [JPERF-298].
 - Upgraded Selenium to 3.141.59